:sparkles:Implement CountLong DP
- Implemented CountLong Datapoint
- removed Helper.h
- Moved callback typedefs to Datapoint.h
- Various typos
